+++ view/frontend/web/js/components/steps/TheShippingStep.vue
@@ -5,19 +5,11 @@
   >
     <h2>{{ $t('shipping.title') }}</h2>
     <form class="shipping-address__form">
-      <BaseInput
-        v-model.trim="$v.customer.email.$model"
-        :validation="$v.customer.email"
-        label="Email"
-        name="email"
-        type="email"
-        @input="checkIsEmailAvailable"
+      <CustomerEmailField
+        v-if="!isCustomerLoggedIn"
+        ref="customerEmail"
+        @ready="isReady => customerEmailReadyToSubmit = isReady"
       />
-      <span
-        v-if="emailAvailabilityMessage"
-        v-html="emailAvailabilityMessage"
-      />
-      <hr>
       <div>
         <BaseInput
           v-model="$v.address.firstname.$model"
@@ -138,16 +130,17 @@
 <script>
 import BaseButton from '../BaseButton.vue'
 import BaseInput from '../BaseInput.vue'
+import CustomerEmailField from '../CustomerEmailField.vue'
 import ShippingMethods from '../ShippingMethods.vue'
 import Multiselect from 'vue-multiselect'
-import axios from 'axios'
 import countries from '../../data/countries.json'
-import { required, email, requiredIf } from 'vuelidate/lib/validators'
+import { required, requiredIf } from 'vuelidate/lib/validators'
 
 export default {
   components: {
     BaseButton,
     BaseInput,
+    CustomerEmailField,
     Multiselect,
     ShippingMethods
   },
@@ -166,22 +159,13 @@ export default {
         region_id: '',
         company: ''
       },
-      customer: {
-        email: '',
-        emailAvailable: false
-      },
       countries,
+      customerEmailReadyToSubmit: false,
       shippingMethodsReadyToSubmit: false,
       loader: false
     }
   },
   validations: {
-    customer: {
-      email: {
-        required,
-        email
-      }
-    },
     address: {
       firstname: {
         required
@@ -217,67 +201,47 @@ export default {
     }
   },
   computed: {
-    baseUrl () {
-      return this.$store.state.baseUrl
-    },
     step () {
       return this.$store.state.step
     },
-    emailAvailabilityMessage () {
-      if (this.customer.email !== '' && !this.$v.customer.email.$error) {
-        if (this.customer.emailAvailable) {
-          return this.$t('shipping.createAccount')
-        } else {
-          return this.$t('shipping.logIn', { link: this.loginUrl })
-        }
-      } else {
-        return false
-      }
+    isCustomerLoggedIn () {
+      return this.$store.getters.isCustomerLoggedIn
     },
     regions () {
       return this.$store.getters.regionsByCountryId(this.address.country_id.value)
-    },
-    loginUrl () {
-      return this.baseUrl + 'customer/account/login/'
     }
   },
   methods: {
-    checkIsEmailAvailable () {
-      const options = {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        data: JSON.stringify({
-          'customerEmail': this.customer.email
-        }),
-        url: `${this.baseUrl}rest/V1/customers/isEmailAvailable`
-      }
-      axios(options)
-        .then(({data}) => {
-          this.customer.emailAvailable = data
-        })
-        .catch(error => {
-          console.error(`${this.$t('error.generalError')} \n`, error)
-        })
-    },
     onCountryChange () {
       this.$store.dispatch('updateShippingMethods', this.address.country_id.value)
     },
     goToNextStep () {
+      if (!this.isCustomerLoggedIn) {
+        this.$refs.customerEmail.touch()
+      }
+
       this.$refs.shippingsMethods.touch()
       this.$v.$touch()
-      if (this.$v.$invalid || !this.shippingMethodsReadyToSubmit) {
+
+      if (
+        this.$v.$invalid ||
+        !this.shippingMethodsReadyToSubmit ||
+        (!this.isCustomerLoggedIn && !this.customerEmailReadyToSubmit)
+      ) {
         return
       }
 
       this.loader = true
-      this.$store.commit('setCustomerEmail', this.customer.email)
+
       this.$store.commit(
         'setAddress',
         { type: 'shippingAddress', address: this.address }
       )
+
       this.$store.dispatch('setShippinInformation').then(() => {
         this.loader = false
       })
+
       this.$store.dispatch('getTotals')
     }
   }